[Move] Commands and interface logic to `/ui`
**Major**:
- The logic for commands and everything leveraging `prompt_toolkit` (including the pretty printing) was in `/core/chat`
- It's not a huge problem in itself but it means we were mixing frontend code with backend code.
- So I finally took the time to move it to the `/ui` folder as part of the current cleanup session.
