+++ hatchling/ui/hatch_commands.py
@@ -11,14 +11,13 @@ from pathlib import Path
 from prompt_toolkit import print_formatted_text
 from prompt_toolkit.formatted_text import FormattedText
 
-from hatchling.core.chat.abstract_commands import AbstractCommands
-from hatchling.mcp_utils.manager import mcp_manager
-
 # Import Hatch components - assumes Hatch is installed or available in the Python path
 # from hatch import HatchEnvironmentManager
 from hatch import create_package_template
 
 from hatchling.config.i18n import translate
+from hatchling.mcp_utils.manager import mcp_manager
+from hatchling.ui.abstract_commands import AbstractCommands
 
 class HatchCommands(AbstractCommands):
     """Handles Hatch package manager commands in the chat interface."""
