+++ hatchling/ui/model_commands.py
@@ -7,7 +7,7 @@ Commands follow the format 'llm:target:action' for clarity and consistency.
 from prompt_toolkit import print_formatted_text
 from prompt_toolkit.formatted_text import FormattedText
 
-from hatchling.core.chat.abstract_commands import AbstractCommands
+from hatchling.ui.abstract_commands import AbstractCommands
 from hatchling.core.llm.model_manager_api import ModelManagerAPI
 from hatchling.config.llm_settings import LLMSettings
