+++ hatchling/ui/settings_commands.py
@@ -6,19 +6,16 @@ management through the chat interface.
 """
 
 import json
-import sys
 from pathlib import Path
-from typing import Dict, Any, Optional, List
+from typing import Dict, Any, List
 
-import tomli_w as toml_write
-import tomli as toml_read
 import yaml
 
 from prompt_toolkit import PromptSession, print_formatted_text
 from prompt_toolkit.formatted_text import FormattedText
 
-from hatchling.core.chat.abstract_commands import AbstractCommands
 from hatchling.config.i18n import translate
+from hatchling.ui.abstract_commands import AbstractCommands
 
 
 class SettingsCommands(AbstractCommands):
